requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
python-dotenv==1.0.0
supabase==2.3.4
requests==2.31.0
